move mptable & mpspec to vm-vcpu-ref
This is the first step required for upstreaming this functionality.
