+++ src/vm-vcpu-ref/Cargo.toml
@@ -13,6 +13,7 @@ keywords = ["virt", "kvm", "vm"]
 kvm-ioctls = { version = "0.8.0" }
 kvm-bindings = { version = "0.4.0", features = ["fam-wrappers"] }
 vm-memory = "0.6.0"
+libc = "0.2.76"
 
 [dev-dependencies]
 vm-memory = { version = "0.6.0", features = ["backend-mmap"] }
