+++ src/vm-vcpu-ref/coverage_config_x86_64.json
@@ -1,5 +1,5 @@
 {
-    "coverage_score": 94.8,
+    "coverage_score": 90.5,
     "exclude_path": "",
     "crate_features": ""
 }
