+++ src/vm-vcpu-ref/src/x86_64/mod.rs
@@ -5,3 +5,5 @@
 pub mod cpuid;
 /// Abstractions for building a Global Descriptor Table (GDT).
 pub mod gdt;
+pub mod mpspec;
+pub mod mptable;
