+++ src/vm-vcpu-ref/src/x86_64/mpspec.rs
